Test db settings are now read from config.

+++ end2end/tests/fixtures/database.ts
@@ -1,31 +1,43 @@
 import {test as base} from '@playwright/test';
 import {Client} from 'pg'
+import * as fs from "node:fs/promises";
+
 
 export class DatabaseHelper {
-    private readonly client: Client;
+    private client: Client | undefined;
     private readonly workerId: number;
 
     constructor(id: number) {
         this.workerId = id;
-        this.client = new Client({
-            user: 'lsf',
-            password: 'lsf',
-            host: 'localhost',
-            database: 'lsf_test',
-            port: 5432,
-        });
+
     }
 
     async connect() {
+        const configFile = await fs.readFile('../config/configuration.test.json');
+        const config: any = JSON.parse(configFile.toString());
+        const dbConfig: any = config.database;
+        this.client = new Client({
+            user: dbConfig.username,
+            password: dbConfig.password,
+            host: dbConfig.host,
+            database: dbConfig.database_name,
+            port: dbConfig.port,
+        });
         await this.client.connect();
     }
 
     async query(sql: string, params: any[] = []) {
+        if (!this.client) {
+            throw new Error("Call connect() before using the client.");
+        }
         const result = await this.client.query(sql, params);
         return result.rows;
     }
 
     async addTestUser(lang: string) {
+        if (!this.client) {
+            throw new Error("Call connect() before using the client.");
+        }
         const username = lang + "_testuser_" + this.workerId;
         // 'password' hashed
         const hash = "$2a$12$2W3AcX2RnI3ZJSwrvWbar.x6FL.nK63niONl.d.mv39bTG5Ru/E9G";
@@ -38,6 +50,9 @@ export class DatabaseHelper {
     }
 
     async deleteTestUser(username: string) {
+        if (!this.client) {
+            throw new Error("Call connect() before using the client.");
+        }
         const query = "DELETE FROM account WHERE username = $1";
         await this.client.query(query, [username]);
     }
